More improvements to physics update

+++ src/physics/MultiRayShape.cc
@@ -101,9 +101,9 @@ void MultiRayShape::Init()
   maxRange = this->rangeElem->GetValueDouble("max");
 
   // Create and array of ray collisions
-  for (unsigned int j = 0; j < vertSamples; j++)
+  for (unsigned int j = 0; j < (unsigned int)vertSamples; j++)
   {
-    for (unsigned int i = 0; i < horzSamples; i++)
+    for (unsigned int i = 0; i < (unsigned int)horzSamples; i++)
     {
       yawAngle = (horzSamples == 1) ? 0 : 
         i * yDiff / (horzSamples - 1) + horzMinAngle;
